feat: [ON-3245] move roles to types.ts
