+++ app/src/i18n/locales/de/settings.json
@@ -23,7 +23,7 @@
   "all": "Hinzufügen",
   "role": "Rolle",
   "admin": "Administrator",
-  "contributor": "Mitwirkender",
+  "user": "Mitwirkender",
   "save-changes": "Änderungen speichern",
   "my-files-sub-title": "Hier finden Sie alle Ihre hochgeladenen Dateien, ausstehende Anfragen zur Harmonisierung und im Treibhausgasinventar enthaltene Dateien",
   "all-inventory-years": "alle Inventarjahre",
