+++ app/src/i18n/locales/en/settings.json
@@ -23,7 +23,7 @@
   "email": "Email",
   "role": "Role",
   "admin": "Admin",
-  "contributor": "Contributor",
+  "user": "Contributor",
   "save-changes": "Save Changes",
   "my-files-sub-title": "Here you can find all your files uploaded, pending request harmonization, and included in the GHG Inventory",
   "all-inventory-years": "all inventory years",
