+++ app/src/i18n/locales/es/settings.json
@@ -19,7 +19,7 @@
   "state-province": "Estado / Provincia",
   "country": "País",
   "admin": "Administrador",
-  "contributor": "Colaborador",
+  "user": "Colaborador",
   "name": "Nombre",
   "all": "Todos",
   "select": "Seleccionar",
