+++ app/src/i18n/locales/pt/settings.json
@@ -23,7 +23,7 @@
   "email": "E-mail",
   "role": "Função",
   "admin": "Administrador",
-  "contributor": "Contribuinte",
+  "user": "Contribuinte",
   "save-changes": "Salvar Alterações",
   "my-files-sub-title": "Aqui você pode encontrar todos os seus arquivos enviados, solicitações pendentes de harmonização e incluídos no Inventário de GEE",
   "all-inventory-years": "todos os anos de inventário",
